+++ actix-rt/CHANGES.md
@@ -1,9 +1,11 @@
 # Changes
 
 ## Unreleased - 2021-xx-xx
+* Add `io-uring` feature for enabling async file I/O on linux. [#374]
 * The `spawn` method can now resolve with non-unit outputs. [#369]
 
 [#369]: https://github.com/actix/actix-net/pull/369
+[#374]: https://github.com/actix/actix-net/pull/374
 
 
 ## 2.2.0 - 2021-03-29